requirements.txt
Django==6.0
django-cors-headers==4.9.0
django_environ==0.12.0
djangorestframework==3.16.1
djoser==2.3.3
drf-spectacular==0.28.0
drf_standardized_errors==0.15.0
factory_boy==3.3.3
Faker==40.13.0
gunicorn==23.0.0
hiredis==3.3.1
psycopg[binary]==3.2.12
pytest==9.0.3
pytest-cov==7.1.0
pytest-django==4.12.0
pytest-html==4.2.0
pytest-mock==3.15.1
pytest-sugar==1.1.1
redis==7.3.0
